+++ noodles-bam/examples/bam_filter_data.rs
@@ -2,7 +2,7 @@
 //!
 //! That is, there is a single alignment hit count (SAM record data tag `NM` = 1).
 
-use std::{env, fs::File, io};
+use std::{env, io};
 
 use noodles_bam as bam;
 use noodles_sam::{alignment::Record, record::data::field::Tag};
@@ -23,7 +23,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut args = env::args().skip(1);
     let src = args.next().expect("missing src");
 
-    let mut reader = File::open(src).map(bam::Reader::new)?;
+    let mut reader = bam::reader::Builder::default().build_from_path(src)?;
     let header = reader.read_header()?.parse()?;
     reader.read_reference_sequences()?;