+++ noodles-bam/examples/bam_reheader.rs
@@ -4,7 +4,7 @@
 //!
 //! Verify the output by piping to `samtools view --no-PG --with-header`.
 
-use std::{env, fs::File, io};
+use std::{env, io};
 
 use noodles_bam as bam;
 use noodles_sam as sam;
@@ -12,7 +12,7 @@ use noodles_sam as sam;
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = File::open(src).map(bam::Reader::new)?;
+    let mut reader = bam::reader::Builder::default().build_from_path(src)?;
     let mut header: sam::Header = reader.read_header()?.parse()?;
     reader.read_reference_sequences()?;