+++ noodles-bam/examples/bam_reset.rs
@@ -5,7 +5,6 @@
 
 use std::{
     env,
-    fs::File,
     io::{self, BufWriter},
 };
 
@@ -20,7 +19,7 @@ use noodles_sam::{
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = File::open(src).map(bam::Reader::new)?;
+    let mut reader = bam::reader::Builder::default().build_from_path(src)?;
     let header: sam::Header = reader.read_header()?.parse()?;
     reader.read_reference_sequences()?;