+++ noodles-bam/examples/bam_view.rs
@@ -2,7 +2,7 @@
 //!
 //! The result matches the output of `samtools view <src>`.
 
-use std::{env, fs::File, io};
+use std::{env, io};
 
 use noodles_bam as bam;
 use noodles_sam::{self as sam, AlignmentWriter};
@@ -10,7 +10,7 @@ use noodles_sam::{self as sam, AlignmentWriter};
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
-    let mut reader = File::open(src).map(bam::Reader::new)?;
+    let mut reader = bam::reader::Builder::default().build_from_path(src)?;
     let header = reader.read_header()?.parse()?;
     reader.read_reference_sequences()?;